pyproject.toml
[project]
name = "mcp-server-redis"
version = "0.1.0"
description = "MCP server to interact with redis server, aws memory DB, etc for caching or other use-cases where in-memory and key-value based storage is appropriate"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.2.1",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
]
